requirements.txt
python-telegram-bot==20.7
aiohttp>=3.9.0
Pillow>=10.3.0
